refactor(state): split store.go into per-domain files
Carves the 3386-LOC umbrella into 14 domain-scoped files: runs,
events, artifacts, cancellation, metrics, progress, chat, ontology,
forks, retros, decisions, webhooks, outcomes, orchestration. store.go
now retains only types, constants, the StateStore interface, the
constructor, migration init, pipeline-state methods, and Close.

Pure file decomposition: no signature changes, no interface changes,
no behaviour change. StateStore umbrella retained — consumer
migration to narrow interfaces deferred to Part 2.

Partial of #1496. Closes the file-split portion.
